ref(FormEndPage): extract EndPageBlock component
Basically the same thing, but to encapsulate component for now to accommodate feedback form component
