+++ frontend/src/features/public-form/components/FormEndPage/components/EndPageBlock/ThankYouSvgr.tsx
