+++ frontend/src/features/public-form/components/FormEndPage/components/EndPageBlock/index.ts
@@ -1,0 +1,1 @@
+export * from './EndPageBlock'
